pyproject.toml
[project]
name = "teri-teri"
description = "Python discord bot for exploring Houkai Gakuen 2 equipdex"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "asyncpg<0.31.0,>=0.30.0",
    "discord-py<3.0.0,>=2.4.0",
    "hg2-data-extractor<0.6.0,>=0.5.1",
    "hg2-item-parser<0.8.0,>=0.7.1",
    "pydantic-settings<3.0.0,>=2.6.1",
    "sqlalchemy<3.0.0,>=2.0.36",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.12.0,<2.0.0",
    "pre-commit>=4.0.1,<5.0.0",
    "ruff>=0.6.9,<1.0.0",
    "types-passlib>=1.7.7.20240819,<2.0.0",
]

[tool.mypy]
strict = true
ignore_missing_imports = true
exclude = [".venv"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N", # pep8-naming
    "UP",  # pyupgrade
    "TRY", # tryceratops
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "EM", # flake8-errmsg
    "FA", # flake8-future-annotations
    "PT", # flake8-pytest-style
    "T20", # flake8-print
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "ANN", # flake8-annotations
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "ERA001", # commented-out-code
    "RUF001", # ambiguous-unicode-character-string
    "RUF003", # ambiguous-unicode-character-comment
]
ignore = [
    "ANN204", # missing-return-type-special-method (dunder methods)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
